[CausalLM] Refactor PerformanceMetrics: separate from API, add embedding support
- Create models/types.h with PerformanceMetrics struct and getPeakMemoryKb()
- Move performance_metrics/hasRun/getPerformanceMetrics to Transformer base
- Remove causal_lm_api.h dependency from causal_lm.h (fix inverted dependency)
- Add metrics collection to SentenceTransformer::encode()
- Simplify getPerformanceMetrics API to work with both CausalLM and embedding
