+++ Applications/CausalLM/models/causal_lm.h
@@ -39,7 +39,6 @@
 #define WCHAR_P std::string &
 #endif
 
-#include "../api/causal_lm_api.h"
 #include <transformer.h>
 
 namespace causallm {
@@ -80,18 +79,6 @@ public:
    * @return Generated text string
    */
   std::string getOutput(int batch_idx = 0) const;
-
-  /**
-   * @brief Get PerformanceMetrics
-   */
-  PerformanceMetrics getPerformanceMetrics() const {
-    return performance_metrics;
-  }
-
-  /**
-   * @brief get the status of run
-   */
-  bool hasRun() const { return has_run_; }
 
 protected:
   /**
@@ -159,10 +146,6 @@ protected:
   bool SAVE_KVCACHE;
   bool USE_KVCACHE;
   unsigned int global_token_len;
-
-  // Performance metrics
-  PerformanceMetrics performance_metrics;
-  bool has_run_ = false;
 
   std::mt19937 rng; /**< Random Number Gen */
 };
